Show product in lesson page nav bar
