+++ react-components/src/components/stylesheets/Modal.scss
