+++ react-components/src/views/SignupModal/stylesheets/Modal.scss
@@ -1,5 +1,4 @@
-@import '@src/stylesheets/ModalCentreScreen';
-
+@import 'helpers/colours';
 
 // the close button opens inside the modal. in the DOM the modal is not a child element of the react component
 .great-mvp-close {
